requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker  may cause unexpected issues

azure-functions
azure-cosmos==4.5.1
azure-storage-blob==12.18.2

